+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Features
 * require user to confirm password before disabling encryption
 * when enabling encryption from settings page, display a second password confirmation field to avoid mistyping password.
+* allow user to change the current unit price when editing a tradeable asset.
 
 ## 1.0.0-beta (2019-12-10)